[IMP] shopinvader: one binding by lang
