+++ shopinvader/views/product_view.xml
@@ -12,11 +12,13 @@
                 <field name="shopinvader_bind_ids">
                     <tree>
                         <field name="backend_id"/>
+                        <field name="lang_id"/>
                         <field name="sync_date"/>
                     </tree>
                     <form>
                         <group name="backend">
                             <field name="backend_id" widget="selection"/>
+                            <field name="lang_id" widget="selection"/>
                             <field name="seo_title"/>
                             <field name="meta_keywords"/>
                             <field name="meta_description"/>
