+++ shopinvader/views/shopinvader_variant.xml
@@ -11,9 +11,11 @@
             <search string="Products">
                 <field name="name" string="Product"/>
                 <field name="backend_id" string="Shopinvader Backend"/>
+                <field name="lang_id" string="Lang"/>
                 <separator/>
                 <group expand="0" string="Group By">
                     <filter string="Shopinvader Backend" name="group_by_backend_id" domain="[]" context="{'group_by':'backend_id'}"/>
+                    <filter string="Lang" name="group_by_lang_id" domain="[]" context="{'group_by':'lang_id'}"/>
                 </group>
             </search>
         </field>
@@ -25,6 +27,7 @@
         <field name="arch" type="xml">
              <tree string="Product" create="0">
                 <field name="backend_id"/>
+                <field name="lang_id"/>
                 <field name="default_code"/>
                 <field name="name"/>
              </tree>
