+++ shopinvader_search_engine/models/product.py
@@ -38,4 +38,5 @@ class ShopinvaderVariant(models.Model):
                 record.index_id = self.env['se.index'].search([
                     ('backend_id', '=', se_backend.id),
                     ('model_id.model', '=', record._name),
+                    ('lang_id', '=', record.lang_id.id)
                     ])
